+++ packages/widgetbook_generator/test/src/code/widgetbook_folder_instance_test.dart
@@ -41,14 +41,12 @@ void main() {
           WidgetbookFolder(
             name: 'root',
             children: [
-              WidgetbookComponent(
+              WidgetbookLeafComponent(
                 name: 'Component',
-                useCases: [
-                  WidgetbookUseCase(
-                    name: 'Default',
-                    builder: defaultUseCase,
-                  )
-                ],
+                useCase: WidgetbookUseCase(
+                  name: 'Default',
+                  builder: defaultUseCase,
+                ),
               ),
               WidgetbookFolder(
                 name: 'Folder1',